Revert "Configure Louis chatbot for the Python SICP textbook (#4085)"
This reverts commit 4ba2edf6a8c5e5fc2d7b5410ad8db4b876b5cde2.
