+++ src/commons/application/types/SessionTypes.ts
@@ -38,11 +38,9 @@ export type SessionState = {
   readonly enableGame?: boolean;
   readonly enableAchievements?: boolean;
   readonly enableLlmGrading?: boolean;
-  readonly enableLouisChatbot?: boolean;
   readonly llmModel?: string;
   readonly llmApiUrl?: string;
   readonly llmCourseLevelPrompt?: string;
-  readonly louisChatbotPrompt?: string;
   readonly pixelbotRoutingPrompt?: string;
   readonly pixelbotAnswerPrompt?: string;
   readonly feedbackUrl?: string;
@@ -115,7 +113,6 @@ export type CourseConfiguration = {
   enableGame: boolean;
   enableAchievements: boolean;
   enableLlmGrading?: boolean;
-  enableLouisChatbot?: boolean;
   enableOverallLeaderboard: boolean;
   enableContestLeaderboard: boolean;
   topLeaderboardDisplay: number;
@@ -127,7 +124,6 @@ export type CourseConfiguration = {
   llmModel?: string;
   llmApiUrl?: string;
   llmCourseLevelPrompt?: string;
-  louisChatbotPrompt?: string;
   pixelbotRoutingPrompt?: string;
   pixelbotAnswerPrompt?: string;
   feedbackUrl?: string;
