+++ src/new_routes/courses/[courseId]/adminpanel.tsx
@@ -24,13 +24,11 @@ const defaultCourseConfig: UpdateCourseConfiguration = {
   topLeaderboardDisplay: 100,
   topContestLeaderboardDisplay: 10,
   enableLlmGrading: false,
-  enableLouisChatbot: false,
   moduleHelpText: '',
   llmApiKey: '',
   llmModel: '',
   llmApiUrl: '',
   llmCourseLevelPrompt: '',
-  louisChatbotPrompt: '',
   pixelbotRoutingPrompt: '',
   pixelbotAnswerPrompt: '',
   feedbackUrl: '',
@@ -65,12 +63,10 @@ function AdminPanel() {
       topLeaderboardDisplay: session.topLeaderboardDisplay,
       topContestLeaderboardDisplay: session.topContestLeaderboardDisplay,
       enableLlmGrading: session.enableLlmGrading,
-      enableLouisChatbot: session.enableLouisChatbot,
       moduleHelpText: session.moduleHelpText,
       llmModel: session.llmModel,
       llmApiUrl: session.llmApiUrl,
       llmCourseLevelPrompt: session.llmCourseLevelPrompt,
-      louisChatbotPrompt: session.louisChatbotPrompt,
       pixelbotRoutingPrompt: session.pixelbotRoutingPrompt,
       pixelbotAnswerPrompt: session.pixelbotAnswerPrompt,
       feedbackUrl: session.feedbackUrl,
@@ -85,9 +81,7 @@ function AdminPanel() {
     session.topContestLeaderboardDisplay,
     session.enableGame,
     session.enableLlmGrading,
-    session.enableLouisChatbot,
     session.moduleHelpText,
-    session.louisChatbotPrompt,
     session.viewable,
     session.llmModel,
     session.llmApiUrl,
