+++ src/new_routes/sicpy/_layout.tsx
@@ -5,9 +5,6 @@ import classNames from 'classnames';
 import { useEffect, useRef } from 'react';
 import { Outlet, useLocation, useNavigate, useParams } from 'react-router';
 import { readLocalStorage } from 'src/commons/hooks/useLocalStorageState';
-import Constants from 'src/commons/utils/Constants';
-import { useSession } from 'src/commons/utils/Hooks';
-import type { SicpSection } from 'src/features/sicp/chatCompletion/chatCompletion';
 import { CodeSnippetProvider } from 'src/features/sicp/CodeSnippetProvider';
 import { ParseJsonError } from 'src/features/sicp/parser/ParseJson';
 import { scrollRefIntoView } from 'src/features/sicp/utils/SicpUtils';
@@ -16,48 +13,17 @@ import {
   SICPY_INDEX,
   useSicPySectionQuery,
 } from 'src/features/textbook/hooks/useTextbookSectionQuery';
-import Chatbot from 'src/pages/sicp/subcomponents/chatbot/Chatbot';
 
 import SicpErrorBoundary from '../../features/sicp/errors/SicpErrorBoundary';
 import getSicpError, { SicpErrorType } from '../../features/sicp/errors/SicpErrors';
 
 const loadingComponent = <NonIdealState title="Loading Content" icon={<Spinner />} />;
 
-const getText = () => {
-  const divs = document.querySelectorAll('p.sicp-text');
-  let visibleParagraphs = '';
-
-  divs.forEach(div => {
-    const rect = div.getBoundingClientRect();
-
-    if (
-      rect.top <= window.innerHeight &&
-      rect.bottom >= 0 &&
-      rect.left <= window.innerWidth &&
-      rect.right >= 0
-    ) {
-      const text = div.textContent;
-      if (text) {
-        visibleParagraphs += text + '\n';
-      }
-    }
-  });
-
-  return visibleParagraphs;
-};
-
 function SicPyLayout() {
-  const { section } = useParams();
+  const { section } = useParams<{ section: string }>();
   const parentRef = useRef<HTMLDivElement>(null);
   const navigate = useNavigate();
   const location = useLocation();
-
-  const { isLoggedIn, enableLouisChatbot } = useSession();
-
-  function getSection(): SicpSection {
-    // To discard the '/sicpy/'
-    return location.pathname.replace('/sicpy/', '') as SicpSection;
-  }
 
   // Handle rerouting to the latest viewed section when clicking from the main
   // application navbar. Navigate replace logic is used to allow the user to
@@ -107,9 +73,6 @@ function SicPyLayout() {
           )}
         </CodeSnippetProvider>
       </SicpErrorBoundary>
-      {isLoggedIn && enableLouisChatbot && Constants.featureFlags.enableSicpChatbot && (
-        <Chatbot getSection={getSection} getText={getText} />
-      )}
     </div>
   );
 }
